Add transitive mode
