+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/classifier/Classifier.java
@@ -11,7 +11,7 @@ import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import edu.kit.kastel.sdq.lissa.ratlr.Configuration;
+import edu.kit.kastel.sdq.lissa.ratlr.configuration.ModuleConfiguration;
 import edu.kit.kastel.sdq.lissa.ratlr.elementstore.ElementStore;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.Element;
 
@@ -60,7 +60,7 @@ public abstract class Classifier {
 
     protected abstract Classifier copyOf();
 
-    public static Classifier createClassifier(Configuration.ModuleConfiguration configuration) {
+    public static Classifier createClassifier(ModuleConfiguration configuration) {
         return switch (configuration.name().split(CONFIG_NAME_SEPARATOR)[0]) {
             case "mock" -> new MockClassifier();
             case "simple" -> new SimpleClassifier(configuration);
