+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/embeddingcreator/EmbeddingCreator.java
@@ -2,7 +2,7 @@ package edu.kit.kastel.sdq.lissa.ratlr.embeddingcreator;
 
 import java.util.List;
 
-import edu.kit.kastel.sdq.lissa.ratlr.Configuration;
+import edu.kit.kastel.sdq.lissa.ratlr.configuration.ModuleConfiguration;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.Element;
 
 public abstract class EmbeddingCreator {
@@ -12,7 +12,7 @@ public abstract class EmbeddingCreator {
 
     public abstract List<float[]> calculateEmbeddings(List<Element> elements);
 
-    public static EmbeddingCreator createEmbeddingCreator(Configuration.ModuleConfiguration configuration) {
+    public static EmbeddingCreator createEmbeddingCreator(ModuleConfiguration configuration) {
         return switch (configuration.name()) {
             case "ollama" -> new OllamaEmbeddingCreator(configuration);
             case "openai" -> new OpenAiEmbeddingCreator(configuration);
