+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/postprocessor/IdentityPostprocessor.java
@@ -3,14 +3,14 @@ package edu.kit.kastel.sdq.lissa.ratlr.postprocessor;
 import java.util.HashSet;
 import java.util.Set;
 
-import edu.kit.kastel.sdq.lissa.ratlr.Configuration;
+import edu.kit.kastel.sdq.lissa.ratlr.configuration.ModuleConfiguration;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.TraceLink;
 
 public class IdentityPostprocessor extends TraceLinkIdPostprocessor {
 
     private final boolean reverse;
 
-    public IdentityPostprocessor(Configuration.ModuleConfiguration moduleConfiguration) {
+    public IdentityPostprocessor(ModuleConfiguration moduleConfiguration) {
         this.reverse = moduleConfiguration.argumentAsBoolean("reverse", false);
     }
