+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/preprocessor/CodeChunkingPreprocessor.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import edu.kit.kastel.sdq.lissa.ratlr.Configuration;
+import edu.kit.kastel.sdq.lissa.ratlr.configuration.ModuleConfiguration;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.Artifact;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.Element;
 
@@ -22,7 +22,7 @@ public class CodeChunkingPreprocessor extends Preprocessor {
     private final List<RecursiveSplitter.Language> languages;
     private final int chunkSize;
 
-    public CodeChunkingPreprocessor(Configuration.ModuleConfiguration configuration) {
+    public CodeChunkingPreprocessor(ModuleConfiguration configuration) {
         this.languages = Arrays.stream(
                         configuration.argumentAsString("language").split(","))
                 .map(RecursiveSplitter.Language::valueOf)
