+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/resultaggregator/ResultAggregator.java
@@ -3,8 +3,8 @@ package edu.kit.kastel.sdq.lissa.ratlr.resultaggregator;
 import java.util.List;
 import java.util.Set;
 
-import edu.kit.kastel.sdq.lissa.ratlr.Configuration;
 import edu.kit.kastel.sdq.lissa.ratlr.classifier.ClassificationResult;
+import edu.kit.kastel.sdq.lissa.ratlr.configuration.ModuleConfiguration;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.Element;
 import edu.kit.kastel.sdq.lissa.ratlr.knowledge.TraceLink;
 
@@ -14,7 +14,7 @@ public abstract class ResultAggregator {
             List<Element> targetElements,
             List<ClassificationResult> classificationResults);
 
-    public static ResultAggregator createResultAggregator(Configuration.ModuleConfiguration configuration) {
+    public static ResultAggregator createResultAggregator(ModuleConfiguration configuration) {
         return switch (configuration.name()) {
             case "any_connection" -> new AnyResultAggregator(configuration);
             default -> throw new IllegalStateException("Unexpected value: " + configuration.name());
