+++ src/test/java/edu/kit/kastel/sdq/lissa/ratlr/ArchitectureTest.java
@@ -2,14 +2,18 @@ package edu.kit.kastel.sdq.lissa.ratlr;
 
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.*;
 
+import java.util.List;
 import java.util.UUID;
+import java.util.function.Consumer;
+import java.util.stream.Stream;
 
 import com.tngtech.archunit.junit.AnalyzeClasses;
 import com.tngtech.archunit.junit.ArchTest;
 import com.tngtech.archunit.lang.ArchRule;
+import edu.kit.kastel.sdq.lissa.ratlr.utils.Environment;
 import edu.kit.kastel.sdq.lissa.ratlr.utils.KeyGenerator;
 
-@AnalyzeClasses(packages = "edu.kit.kastel.sdq.lissa.ratlr")
+@AnalyzeClasses(packages = "edu.kit.kastel.sdq.lissa")
 class ArchitectureTest {
 
     @ArchTest
@@ -28,4 +32,16 @@ class ArchitectureTest {
             .should()
             .accessClassesThat()
             .haveNameMatching(UUID.class.getName());
+
+    @ArchTest
+    static final ArchRule noForEachInCollectionsOrStream = noClasses()
+            .should()
+            .callMethod(Stream.class, "forEach", Consumer.class)
+            .orShould()
+            .callMethod(Stream.class, "forEachOrdered", Consumer.class)
+            .orShould()
+            .callMethod(List.class, "forEach", Consumer.class)
+            .orShould()
+            .callMethod(List.class, "forEachOrdered", Consumer.class)
+            .because("Lambdas should be functional. ForEach is typically used for side-effects.");
 }
